Fix usernames on chat toggling in and out
The filter function for checking if the username should be shown or not
pulls chats by index to check them against their neighbors. The chats
shown in the UI are pre-filtered by if they are within the last X hours
(hide day old chats by default), so the $index coming off those
pre-filtered chats was being passed into the filter function which was
comparing chats by index against a chat array which wasn't filtered.

This meant that as X 24 hour old chats fell out of view all the chat
usernames were being shown based on whether the username should be shown
of the chat X back, which in back-and-forth chat might look reasonable,
but in large chunks of machine chat looks very odd.

Closes #144